requirements.txt
lxml==5.4.0
python-docx==1.1.2
tangled-up-in-unicode==0.2.0
typing_extensions==4.13.2
